bq/requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-dataform==0.5.11